# reduce set allocation