100% Set&Create benchmark, `uint64` key and `[512]byte` value:

**before:**
```
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkCache/otter_reads=0%,writes=100%-6         	 1000000	      1036 ns/op	    965062 ops/s	     582 B/op	       1 allocs/op

BenchmarkCache/theine_reads=0%,writes=100%-6        	 2075300	       562.5 ns/op	   1777821 ops/s	      76 B/op	       2 allocs/op

BenchmarkCache/ristretto_reads=0%,writes=100%-6     	 6303217	       166.0 ns/op	   6024788 ops/s	     612 B/op	       3 allocs/op
```


**after:**
```
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkCache/otter_reads=0%,writes=100%-6         	 1000000	      1196 ns/op	    836077 ops/s	     582 B/op	       1 allocs/op

BenchmarkCache/theine_reads=0%,writes=100%-6        	 2076139	       559.9 ns/op	   1786018 ops/s	      12 B/op	       0 allocs/op

BenchmarkCache/ristretto_reads=0%,writes=100%-6     	 7765894	       170.9 ns/op	   5851165 ops/s	     612 B/op	       3 allocs/op
```


The reason Ristretto is fast (170.9 ns/op) is that it [proactively drops Set](https://github.com/dgraph-io/ristretto/blob/ce5561f7d9a89363e9d889095428fa9ef84f9c76/cache.go#L303):
```go
select {
	case c.setBuf <- i:
		return true
	default:
		...
```